ImageCombiner.App: moved file list to separate control, reworked file handling, added file list management. ImageCombiner.Core: Added basic frame support

+++ src/ImageCombiner.Core/Models/FramingOptions.cs
@@ -2,10 +2,10 @@
 
 public class FramingOptions
 {
-    public float Thickness { get; }
+    public int Thickness { get; }
     public FrameColor Color { get; }
 
-    public FramingOptions(float thickness, FrameColor color)
+    public FramingOptions(int thickness, FrameColor color)
     {
         Thickness = thickness;
         Color = color;
